+++ src/__tests__/handler/links.test.ts
@@ -2,6 +2,8 @@ import { Neovim } from '@chemzqm/neovim'
 import { Disposable, DocumentLink, Range } from 'vscode-languageserver-protocol'
 import LinksHandler from '../../handler/links'
 import languages from '../../languages'
+import workspace from '../../workspace'
+import events from '../../events'
 import { disposeAll } from '../../util'
 import helper from '../helper'
 
@@ -11,7 +13,7 @@ let disposables: Disposable[] = []
 beforeAll(async () => {
   await helper.setup()
   nvim = helper.nvim
-  links = (helper.plugin as any).handler.links
+  links = helper.plugin.getHandler().links
 })
 
 afterAll(async () => {
@@ -65,6 +67,11 @@ describe('Links', () => {
   it('should open link at current position', async () => {
     await nvim.setLine('foo')
     await nvim.command('normal! 0')
+    disposables.push(workspace.registerTextDocumentContentProvider('test', {
+      provideTextDocumentContent: () => {
+        return 'test'
+      }
+    }))
     disposables.push(languages.registerDocumentLinkProvider([{ language: '*' }], {
       provideDocumentLinks(_doc, _token) {
         return [
@@ -95,5 +102,40 @@ describe('Links', () => {
     }))
     let res = await links.openCurrentLink()
     expect(res).toBe(false)
+  })
+
+  it('should show tooltip', async () => {
+    await nvim.setLine('foo')
+    await nvim.call('cursor', [1, 1])
+    disposables.push(languages.registerDocumentLinkProvider([{ language: '*' }], {
+      provideDocumentLinks(_doc, _token) {
+        let link = DocumentLink.create(Range.create(0, 0, 0, 5))
+        link.tooltip = 'test'
+        return [link]
+      },
+      resolveDocumentLink(link) {
+        link.target = 'http://example.com'
+        return link
+      }
+    }))
+    await links.showTooltip()
+    let win = await helper.getFloat()
+    let buf = await win.buffer
+    let lines = await buf.lines
+    expect(lines[0]).toMatch('test')
+  })
+
+  it('should enable tooltip on CursorHold', async () => {
+    let doc = await workspace.document
+    helper.updateConfiguration('links.tooltip', true)
+    await nvim.setLine('http://www.baidu.com')
+    await nvim.call('cursor', [1, 1])
+    let link = await links.getCurrentLink()
+    expect(link).toBeDefined()
+    await events.fire('CursorHold', [doc.bufnr])
+    let win = await helper.getFloat()
+    let buf = await win.buffer
+    let lines = await buf.lines
+    expect(lines[0]).toMatch('Press')
   })
 })